pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "posit-gskill"
version = "0.1.0a1"
description = "Evolutionary optimization of Claude Skills for R package development using GEPA"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Posit", email = "support@posit.co" }
]
keywords = [
    "claude",
    "gepa",
    "optimization",
    "r-packages",
    "skills",
    "testing",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "gepa",
    "docker",
    "pyyaml",
    "tree-sitter-language-pack",
    "mlflow",
    "pydantic",
    "pydantic-settings",
    "rich",
    "python-dotenv",
    # For PR mining
    "PyGithub>=2.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "ruff",
]

[project.scripts]
posit-gskill = "posit_gskill.cli:main"

[project.urls]
Homepage = "https://github.com/posit-dev/posit-gskill"
Documentation = "https://github.com/posit-dev/posit-gskill#readme"
Repository = "https://github.com/posit-dev/posit-gskill"
Issues = "https://github.com/posit-dev/posit-gskill/issues"

[tool.hatch.build.targets.wheel]
packages = ["posit_gskill", "task_generator", "evaluation", "optimization", "bench"]

# uv configuration
[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "ruff",
    "ty",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "TCH", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ty.environment]
python-version = "3.12"

[tool.ty.rules]
possibly-unresolved-reference = "warn"
invalid-return-type = "warn"
invalid-argument-type = "warn"
unresolved-import = "warn"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration_local: Integration tests requiring Docker (skipped if Docker unavailable)",
    "integration_provider: Integration tests requiring real provider credentials",
]
